pyproject.toml
[project]
name = 'virtual-mechanism'
readme = { file = "README.md", content-type = "text/markdown" }
authors = [
  { name = "Pride Leong", email = "lykling.lyk@gmail.com" }
]
description = "virtual mechanism for mock and test"
requires-python = ">3.6"
dynamic = ["version"]
dependencies = [
    "cantools>=37.2.0",
    "click>=8.0.4",
    "python-can>=3.3.4",
    "pyyaml>=6.0",
]

[project.scripts]
virtual-mechanism = "virtual_mechanism.__main__:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.dynamic]
version = {attr = "virtual_mechanism.VERSION"}

[tool.setuptools.packages]
find = {}

[tool.setuptools.package-dir]
virtual_mechanism = "virtual_mechanism"

[tool.setuptools.package-data]
virtual_mechanism = ["dbc/*.dbc"]

[project.urls]
"Homepage" = "https://github.com/lykling/virtual-mechanism/tree/main"
"Bug Tracker" = "https://github.com/lykling/virtual-mechanism/issues"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
